[PAGOPA-2043] feat: setting Panache logging to DEBUG for tracing
