--- performance-test/src/config.json
@@ -1,7 +1,0 @@
-{
-    "pspId": "60000000001",
-    "organizationId": "15376371009",
-    "env": "dev",
-    "pspBrokerId": "60000000001",
-    "channelId": "15376371009_04"
-}
