--- performance-test/src/helpers/payloads.json
@@ -1,51 +1,0 @@
-{
-    "create_flow_payload": {
-        "fdr": "",
-        "fdrDate": "",
-        "sender": {
-            "type": "LEGAL_PERSON",
-            "id": "SELBIT2B",
-            "pspId": "",
-            "pspName": "Bank",
-            "pspBrokerId": "60000000001",
-            "channelId": "15376371009_04",
-            "password": "PLACEHOLDER"
-        },
-        "receiver": {
-            "id": "APPBIT2B",
-            "organizationId": "15376371009",
-            "organizationName": "Comune di xyz"
-        },
-        "regulation": "SEPA - Bonifico xzy",
-        "regulationDate": "",
-        "bicCodePouringBank": "UNCRITMMXXX",
-        "totPayments": 3,
-        "sumPayments": 0.03
-    },
-
-    "add_payments_payload": {
-        "payments": [{
-            "iuv": "00000000000000a",
-            "iur": "00000000000000a",
-            "index": 1,
-            "pay": 0.01,
-            "payStatus": "EXECUTED",
-            "payDate": "2023-02-03T12:00:30.900000Z"
-          },{
-            "iuv": "00000000000000b",
-            "iur": "00000000000000b",
-            "index": 2,
-            "pay": 0.01,
-            "payStatus": "EXECUTED",
-            "payDate": "2023-02-03T12:00:30.900000Z"
-          },{
-            "iuv": "00000000000000c",
-            "iur": "00000000000000c",
-            "index": 3,
-            "pay": 0.01,
-            "payStatus": "EXECUTED",
-            "payDate": "2023-02-03T12:00:30.900000Z"
-          }
-        ]
-    }
-}
