--- performance-test/src/modules/client.js
@@ -1,9 +1,0 @@
-import http from 'k6/http';
-
-export function example(rootUrl, params) {
-  const url = `${rootUrl}/`
-
-  const payload = {}
-
-  return http.post(url, JSON.stringify(payload), params);
-}
