+++ src/main/resources/application.properties
@@ -90,4 +90,6 @@ api_config_cache.cron.expr=${ADAPTER_API_CONFIG_CACHE_CRON:*/15 * * * * ?}
 %test.quarkus.mockserver.devservices.config-file=src/test/resources/mockserver.properties
 %test.quarkus.mockserver.devservices.config-dir=src/test/resources/mockserver
 %test.quarkus.mockserver.devservices.log=false
+quarkus.log.category."io.quarkus.mongodb.panache.common.runtime".level=DEBUG
+
